Start of API code for MasterControl.
This has not been tested on Python 3.X and may be at risk of failure.

This is the starting point for AMP -> MasterControl integration through use
of a very simple web server. In the end this will be used to allow AMP to alert
MasterControl of new members, log events, or other information that should
help its operation.